# agendamento/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamento de Consultas",
    "version": "1.0.0",
    "description": "Documentação da API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/users/{userId}/doctors": {
      "get": {
        "summary": "Retorna todos os médicos com quem um utilizador teve consultas",
        "description": "Retorna uma lista de todos os médicos com quem um utilizador específico (paciente)\nteve pelo menos uma consulta agendada.\nEste endpoint demonstra a relação Muitos-para-Muitos entre Utilizadores e Médicos\natravés da tabela de Agendamentos.\nAcessível apenas ao utilizador especificado (se autenticado) ou a um administrador.\n",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "O ID do utilizador (paciente) cujos médicos associados devem ser retornados",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de médicos associados ao utilizador encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "O ID do médico"
                      },
                      "name": {
                        "type": "string",
                        "description": "O nome do médico"
                      },
                      "specialty": {
                        "type": "object",
                        "description": "A especialidade do médico",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID da especialidade"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome da especialidade"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado. O token de autenticação está ausente ou inválido."
          },
          "403": {
            "description": "Acesso negado. O utilizador autenticado não tem permissão para aceder a esta lista de médicos."
          },
          "404": {
            "description": "Utilizador não encontrado com o ID fornecido."
          },
          "500": {
            "description": "Erro interno do servidor ao processar a requisição."
          }
        }
      }
    },
...
    "/doctors": {
      "get": {
        "summary": "Retorna uma lista de médicos",
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "Lista de médicos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar médicos"
          }
        }
      },
      "post": {
        "summary": "Cria um novo médico",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "O nome do médico"
                  },
                  "specialty_id": {
                    "type": "integer",
                    "description": "O ID da especialidade do médico"
                  }
                },
                "required": [
                  "name",
                  "specialty_id"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Médico criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou ausentes"
          }
        }
      }
    },
    "/doctors/{id}": {
      "get": {
        "summary": "Retorna um médico pelo ID",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médico encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Médico não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza os dados de um médico",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "specialty_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Médico atualizado com sucesso"
          },
          "400": {
            "description": "Erro ao atualizar médico"
          },
          "404": {
            "description": "Médico não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Elimina um médico pelo ID",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médico removido com sucesso"
          },
          "404": {
            "description": "Médico não encontradp"
          }
        }
      }
    },
    "/doctors/{id}/appointments": {
      "get": {
        "summary": "Retorna todas as consultas e pacientes de um médico específico",
        "description": "Lista todas as consultas agendadas com um médico cujo ID é especificado no caminho,\nincluindo os detalhes dos pacientes associados a cada consulta,\nbem como os detalhes do médico e a sua especialidade.\nEsta rota requer autenticação. A lógica de autorização (ex: apenas admin ou o próprio médico)\ndeve ser implementada no backend.\n",
        "tags": [
          "Doctors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "O ID do médico cujas consultas e pacientes devem ser retornados",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de consultas com pacientes encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "consultaId": {
                        "type": "integer",
                        "description": "O ID da consulta"
                      },
                      "date": {
                        "type": "string",
                        "format": "date",
                        "description": "A data da consulta"
                      },
                      "time": {
                        "type": "string",
                        "format": "time",
                        "description": "A hora da consulta"
                      },
                      "notes": {
                        "type": "string",
                        "description": "Notas ou descrição da consulta (opcional)",
                        "nullable": true
                      },
                      "paciente": {
                        "type": "object",
                        "description": "Os detalhes do paciente (utilizador) associado a esta consulta",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID do paciente"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome do paciente"
                          },
                          "email": {
                            "type": "string",
                            "format": "email",
                            "description": "O email do paciente"
                          }
                        }
                      },
                      "medico": {
                        "type": "object",
                        "description": "Os detalhes do médico associado a esta consulta",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "O ID do médico"
                          },
                          "name": {
                            "type": "string",
                            "description": "O nome do médico"
                          },
                          "specialty": {
                            "type": "object",
                            "description": "A especialidade do médico",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "description": "O ID da especialidade"
                              },
                              "name": {
                                "type": "string",
                                "description": "O nome da especialidade"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado. O token de autenticação está ausente ou inválido."
          },
          "403": {
            "description": "Acesso negado. O utilizador autenticado não tem permissão para aceder a estas consultas (ex: não é admin, nem o médico em questão)."
          },
          "404": {
            "description": "Médico não encontrado com o ID fornecido."
          },
          "500": {
            "description": "Erro interno do servidor ao processar a requisição."
          }
        }
      }
    },
...
  }
}